genai-backend/requirements.txt
chromadb==1.0.16
fastapi==0.116.1
langchain==0.3.27
langchain_community==0.3.27
langchain_core==0.3.74
langchain_google_genai==2.1.9
motor==3.7.1
protobuf==6.31.1
pydantic==2.11.7
pymongo==4.14.0
PyPDF2==3.0.1
python-dotenv==1.1.1
requests==2.32.4
python-multipart==0.0.9
google-generativeai==0.8.5
google-api-python-client==2.178.0
